Created the Cart Controller.

+++ revature-cart-backend/src/main/java/com/revature/cart/model/Product.java
@@ -1,7 +1,5 @@
 package com.revature.cart.model;
 
-import java.math.BigDecimal;
-
 public class Product {
 	
 	private Long id;
@@ -12,14 +10,14 @@ public class Product {
 	private String model;
 	private String image;
 	private Integer quantity;
-	private BigDecimal unitPrice;
+	private Integer unitPrice;
 	private String color;
 	
 	public Product() {
 	}
 
 	public Product(Long id, String category, String brand, String name, String description, String model,
-			String image, Integer quantity, BigDecimal unitPrice, String color) {
+			String image, Integer quantity, Integer unitPrice, String color) {
 		super();
 		this.id = id;
 		this.category = category;
@@ -81,11 +79,11 @@ public class Product {
 		this.quantity = quantity;
 	}
 
-	public BigDecimal getUnitPrice() {
+	public Integer getUnitPrice() {
 		return unitPrice;
 	}
 
-	public void setUnitPrice(BigDecimal unitPrice) {
+	public void setUnitPrice(Integer unitPrice) {
 		this.unitPrice = unitPrice;
 	}
